32-bit cache for up to +159% performance improvement
Massive performance improvement across the board. Seemingly no drawbacks.

+++ helsing/configuration_adv.h
@@ -63,8 +63,8 @@
 	 * on the args and the configuration.
 	 */
 
-	typedef uint64_t digits_t;
-	#define DIGITS_T_MAX UINT64_MAX
+	typedef uint32_t digits_t;
+	#define DIGITS_T_MAX UINT32_MAX
 
 /*
  * Helper Preprocessor Macros
